+++ src/test/java/roomescape/application/london/ReservationServiceTest.java
@@ -1,4 +1,4 @@
-package roomescape.application;
+package roomescape.application.london;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -18,6 +18,8 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
+import roomescape.application.ReservationService;
+import roomescape.application.ReservationTimeService;
 import roomescape.domain.reservation.Reservation;
 import roomescape.domain.reservation.repository.ReservationRepository;
 import roomescape.domain.time.ReservationTime;
