+++ src/test/java/roomescape/application/london/ReservationTimeServiceTest.java
@@ -1,4 +1,4 @@
-package roomescape.application;
+package roomescape.application.london;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -14,6 +14,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import roomescape.application.ReservationTimeService;
 import roomescape.domain.time.ReservationTime;
 import roomescape.domain.time.repository.ReservationTimeRepository;
 import roomescape.dto.reservationtime.ReservationTimeRequest;
